db: add schema file
Normally we'd use goose's migrations to define the schema used by sqlc
for codegen, but in this instance (using a pre-populated existing database),
we need to expose the full schema for sqlc to use.

Once the ergast API is deprecated, let's remove the mysql seed process, make one
last image, and then construct a new postgres image using the full dump from the
seeded image.

1) Create migrations that match the schema file
2) Dump only the data from the schema image
3) insert all the data somehow

Closes #11.